+++ resources/views/product/index.blade.php
@@ -17,8 +17,10 @@
                     <div class="card">
                         <div class="card-header">
                             <h3 class="card-title">Product List</h3>
-                            <a href="{{ route('product.create') }}" class="btn btn-primary btn-sm float-right"> <i
-                                    class="fas fa-plus"></i> Add New Product</a>
+                            @can('product_add')
+                                <a href="{{ route('product.create') }}" class="btn btn-primary btn-sm float-right"> <i
+                                        class="fas fa-plus"></i> Add New Product</a>
+                            @endcan
                         </div>
                         <!-- /.card-header -->
                         <div class="card-body">
@@ -32,7 +34,9 @@
                                         <th>Code</th>
                                         <th>Category</th>
                                         <th>Price</th>
-                                        <th>Action</th>
+                                        @if (Auth::user()->can('product_edit') || Auth::user()->can('product_delete'))
+                                            <th>Action</th>
+                                        @endif
                                     </tr>
                                 </thead>
                                 <tbody>
@@ -56,15 +60,21 @@
                                             <td> {{ $product->code }}</td>
                                             <td> {{ $product->category->name }}</td>
                                             <td> {{ $product->unit_price }}</td>
-                                            <td>
-                                                <a href="{{ route('product.edit', $product->id) }}"
-                                                    class="btn btn-primary btn-sm"><i class="fas fa-edit"></i></a>
-                                                <button type="button"
-                                                    class="btn btn-sm  btn-danger js-bs-tooltip-enabled mx-2"
-                                                    onclick="deleteProduct(this)" data-id="{{ $product->id }}">
-                                                    <i class="fas fa-trash"></i>
-                                                </button>
-                                            </td>
+                                            @if (Auth::user()->can('product_edit') || Auth::user()->can('product_delete'))
+                                                <td>
+                                                    @can('product_edit')
+                                                        <a href="{{ route('product.edit', $product->id) }}"
+                                                            class="btn btn-primary btn-sm"><i class="fas fa-edit"></i></a>
+                                                    @endcan
+                                                    @can('product_delete')
+                                                        <button type="button"
+                                                            class="btn btn-sm  btn-danger js-bs-tooltip-enabled mx-2"
+                                                            onclick="deleteProduct(this)" data-id="{{ $product->id }}">
+                                                            <i class="fas fa-trash"></i>
+                                                        </button>
+                                                    @endcan
+                                                </td>
+                                            @endif
                                         </tr>
 
                                         <!-- Modal for View product -->
@@ -85,39 +95,41 @@
                                                                 <div class="form-group text-center mb-4">
                                                                     <div class="mb-3">
                                                                         @if ($product->image)
-                                                                        <img src="{{ asset($product->image) }}"
-                                                                            id="photo" alt="Product Image"
-                                                                            height="200">
+                                                                            <img src="{{ asset($product->image) }}"
+                                                                                id="photo" alt="Product Image"
+                                                                                height="200">
                                                                         @else
-                                                                        <img src="https://placehold.co/300x300?font=roboto"
-                                                                            id="photo" alt="Product Image"
-                                                                            class="shadow" height="300">
+                                                                            <img src="https://placehold.co/300x300?font=roboto"
+                                                                                id="photo" alt="Product Image"
+                                                                                class="shadow" height="300">
                                                                         @endif
                                                                     </div>
                                                                 </div>
                                                             </div>
                                                             <div class="col-md-12 text-center">
-                                                               <div class="d-flex">
+                                                                <div class="d-flex">
                                                                     <div class="form-group col-md-6">
                                                                         <label class="mb-3 font-weight-bold">Name</label>
                                                                         <p>{{ $product->name }}</p>
                                                                     </div>
                                                                     <div class="form-group col-md-6">
-                                                                        <label class="mb-3 font-weight-bold">Quantity</label>
+                                                                        <label
+                                                                            class="mb-3 font-weight-bold">Quantity</label>
                                                                         <p>{{ $product->quantity }}</p>
                                                                     </div>
-                                                               </div>
-                                                               <div class="d-flex">
+                                                                </div>
+                                                                <div class="d-flex">
                                                                     <div class="form-group col-md-6">
                                                                         <label class="mb-3 font-weight-bold">Code</label>
                                                                         <p>{{ $product->code }}</p>
                                                                     </div>
                                                                     <div class="form-group col-md-6">
-                                                                        <label class="mb-3 font-weight-bold">Category</label>
+                                                                        <label
+                                                                            class="mb-3 font-weight-bold">Category</label>
                                                                         <p>{{ $product->category->name }}</p>
                                                                     </div>
-                                                               </div>
-                                                               <div class="form-group col-md-12">
+                                                                </div>
+                                                                <div class="form-group col-md-12">
                                                                     <label class="mb-3 font-weight-bold">Description</label>
                                                                     <p>{{ $product->description }}</p>
                                                                 </div>
